INITIAL COMMIT
This is the initial commit. There is one rule that detects try/catch
in tests, but only if the TC is in a test file. I need to add more logic
to check only that it happens within an "it" function.

The docs and tests folders have content as placeholder. They currently
contain content for the "semi" rule to act as an exemplar.