NO-TRY-CATCH-FINALLY NEARLY COMPLETE
The README.md file in the root directory now has a supported rules
section with a single supported rule "no-try-catch-finally".

The no-try-catch-finally.md file now has rule specific markdown versus
being the placeholder file with the default semi rule markdown as it
was before.

In no-try-catch-finally.js the "meta" section is now complete. "Schema"
is now an empty object because the rule does not currerntly support
options. "isTestFile" has been renamed to "isThisFileATestFile" and
simplified.

a new file "try-catch-finally.spec.js" has been added to the root "src"
folder so that both a "test" and "spec" file exists for manually
testing each rule.